Deploy task and added comments
